microservices/notification-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
redis==5.0.1
python-jose[cryptography]==3.3.0
pydantic==2.5.2
python-dotenv==1.0.0
